+++ sources/bomb.cpp
@@ -1,6 +1,6 @@
 #include "../headers/bomb.hpp"
 
-Bomb::Bomb(gameDataRef data, int playerId, int lenghtX, int lenghtY, int explodeTime, float timeCreated, sf::Vector2f pos):
+Bomb::Bomb(gameDataRef data, int playerId, int lenghtX, int lenghtY, float explodeTime, float timeCreated, sf::Vector2f pos):
     data(data),
     id(playerId),
     lenghtX(lenghtX),
